custom-embeddings/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pillow==11.2.0
pydantic==2.11.1
pydantic_core==2.33.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
sentence-transformers==4.0.1
setuptools==78.1.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
python-dotenv==1.0.0
